+++ SimulationKit/skopenclvoidfractionunitcell.cpp
@@ -200,7 +200,7 @@ double SKOpenCLVoidFractionUnitCell::computeVoidFraction(std::vector<cl_float>* 
 
   // Final summation with CPU
   double fraction = 0.0;
-  for (int i=0; i<nWorkGroups; i++)
+  for (size_t i=0; i<nWorkGroups; i++)
     fraction += double(sumReduction[i]);
 
   clReleaseMemObject(rawData);
